refactor : 레거시 레이어 정리(api, modules, stores, config)
