+++ app/src/features/github-sync/index.ts
@@ -1,1 +1,2 @@
 export * from './api/githubApi';
+export type { Branch, CommitDetail, FileChange } from '@/entities/repository';
