--- app/src/i18n.ts
@@ -1,33 +1,0 @@
-import i18n from 'i18next';
-import { initReactI18next } from 'react-i18next';
-import LanguageDetector from 'i18next-browser-languagedetector';
-import Backend from 'i18next-http-backend';
-
-i18n
-  .use(Backend)
-  .use(LanguageDetector)
-  .use(initReactI18next)
-  .init({
-    supportedLngs: ['ko', 'en'],
-    fallbackLng: 'en',
-    backend: {
-      loadPath: '/locales/{{lng}}/{{ns}}.json',
-    },
-    ns: ['translation', 'landing'],
-    defaultNS: 'translation',
-    interpolation: {
-      escapeValue: false,
-    },
-    detection: {
-      caches: ['localStorage', 'cookie'],
-    },
-  });
-
-// document.documentElement.lang 갱신
-i18n.on('languageChanged', (lng) => {
-  if (typeof document !== 'undefined') {
-    document.documentElement.lang = lng;
-  }
-});
-
-export default i18n;
