--- app/src/modules/Annotation.ts
@@ -1,2 +1,0 @@
-// 기존 Annotation.js 내용을 TypeScript로 변환
-// 현재 파일이 비어있거나 간단한 경우, 필요에 따라 추가
